minor fix to data conversion

+++ yolo_ros/node.py
@@ -101,10 +101,10 @@ class YoloROS(Node):
         
             for bbox, cls, conf in zip(self.result[0].boxes.xywh, self.result[0].boxes.cls, self.result[0].boxes.conf):
 
-                self.detection_msg.bbx_center_x.append(float(bbox[0]))
-                self.detection_msg.bbx_center_y.append(float(bbox[1]))
-                self.detection_msg.bbx_size_w.append(float(bbox[2]))
-                self.detection_msg.bbx_size_h.append(float(bbox[3]))
+                self.detection_msg.bbx_center_x.append(int(bbox[0]))
+                self.detection_msg.bbx_center_y.append(int(bbox[1]))
+                self.detection_msg.bbx_size_w.append(int(bbox[2]))
+                self.detection_msg.bbx_size_h.append(int(bbox[3]))
                 self.detection_msg.class_id.append(int(cls))
                 self.detection_msg.confidence.append(float(conf))
 
@@ -166,10 +166,10 @@ class YoloROS(Node):
         
             for bbox, cls, conf in zip(self.result[0].boxes.xywh, self.result[0].boxes.cls, self.result[0].boxes.conf):
 
-                self.detection_msg.bbx_center_x.append(float(bbox[0]))
-                self.detection_msg.bbx_center_y.append(float(bbox[1]))
-                self.detection_msg.bbx_size_w.append(float(bbox[2]))
-                self.detection_msg.bbx_size_h.append(float(bbox[3]))
+                self.detection_msg.bbx_center_x.append(int(bbox[0]))
+                self.detection_msg.bbx_center_y.append(int(bbox[1]))
+                self.detection_msg.bbx_size_w.append(int(bbox[2]))
+                self.detection_msg.bbx_size_h.append(int(bbox[3]))
                 self.detection_msg.class_id.append(int(cls))
                 self.detection_msg.confidence.append(float(conf))
